Move 'MediaType' model to base namespace
We use it for both assets and links.
