+++ rgd/stac/models/mediatype.py
